requirements.txt
torch==2.4.0
torchvision==0.19.0
pytorch-lightning==2.0.4
torchmetrics==0.11.4
tqdm==4.65.0
Pillow==9.3.0
